pyproject.toml
[project]
name = "coffe-shop-test"
version = "0.1.0"
description = ""
authors = [
    {name = "Asaf Mirzoev",email = "asafmirzoev1404@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.13,<0.116.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pydantic-settings (>=2.10.0,<3.0.0)",
    "redis (>=6.2.0,<7.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "argon2-cffi (>=25.1.0,<26.0.0)",
    "alembic (>=1.16.2,<2.0.0)",
    "pydantic[email] (>=2.11.7,<3.0.0)",
    "greenlet (>=3.2.3,<4.0.0)",
    "uvicorn (>=0.34.3,<0.35.0)"
]


[tool.black]
line-length = 88
skip-string-normalization = true
skip-magic-trailing-comma = false
target-version = ['py311']
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.pytest_cache
  | \.tox
  | build
  | dist
  | venv
  | .venv
)/
'''


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
